refactor: split logger and engineLogger
