+++ src/app.js
@@ -13,6 +13,7 @@ const cockpitService = require("./services/cockpit");
 const { setCustomNodes } = require("../src/nodes");
 
 const _log = require("./utils/logger");
+const elog = require("./utils/engineLogger");
 const listeners = require("./utils/engineListener");
 const mqtt = require("./services/mqtt");
 const { db } = require("./utils/db");
@@ -21,7 +22,7 @@ const { setPersist } = require("./middlewares/persist");
 
 const startServer = (port) => {
   const engineLogLevel = process.env.ENGINE_LOG_LEVEL || "error";
-
+  elog.startLogger();
   let engine = getEngine();
   if (!engine) {
     engine = new Engine("knex", db, engineLogLevel);
@@ -49,15 +50,14 @@ const startServer = (port) => {
   const app = new Koa();
   const corsOptions = {
     origin: "*",
-    allowMethods: ["GET", "POST", "DELETE"],
+    allowMethods: ["GET", "POST", "DELETE", "PUT", "PATCH"],
     allowHeaders: ["Content-Type", "Authorization", "Accept", "x-duration", "x-secret"],
   };
   app.use(cors(corsOptions));
   app.use(setPersist(db));
 
   app.use(koaLogger(_log.logger));
-  _log.startLogger();
-
+  
   app.use(freeRouter({ corsOptions }).routes());
 
   app.use(
